Update 1Password secret provider to use reference and check for Session Token
